requirements.txt
azure-functions
openai
requests
flask
python-dotenv
flask-restful